Handle typekit packages as non-catkin packages
kdl_typekit and eigen_typekit currently do not follow catkin conventions.
Although catkin_package() is called internally by orocos_generate_package(),
the typekits cannot be used without installation from devel-space because
typekit headers cannot be found in dependent packages.

This patch adds the build_type cmake declaration to package.xml, so that
the package can be built with catkin_make_isolated or catkin_tools, which
install the headers and libraries in the devel-space.